Extract toggling protected layer to utils
