+++ src/components/protected-areas-layers/protected-areas-layers-component.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import MultipleActiveLayers from 'components/multiple-active-layers';
 
 import { usePaintLayer } from 'hooks/esri';
-import { handleLayerRendered, getToggledLayer } from 'utils/layer-manager-utils';
+import { toggleWDPALayer } from 'utils/layer-manager-utils';
 import { WDPALayers, PROTECTED_AREAS_COLOR, COMMUNITY_AREAS_COLOR } from 'constants/protected-areas';
 import { PROTECTED_AREAS_VECTOR_TILE_LAYER, PROTECTED_AREAS_LAYER_GROUP, COMMUNITY_AREAS_LAYER_GROUP, COMMUNITY_AREAS_VECTOR_TILE_LAYER } from 'constants/layers-slugs';
 
@@ -23,20 +23,7 @@ const ProtectedAreasLayers = ({ handleGlobeUpdating, handleLayerToggle, activeLa
   usePaintLayer(CommunityVTLLayer, 'WDPA_poly_Latest/1', COMMUNITY_AREAS_COLOR);
 
   const toggleLayer = (layersPassed, option) => {
-    const layerNotRendered = !activeLayers.some(layer => layer.title === option.id);
-
-    const layerToggled = getToggledLayer(layers.items, option);
-    
-    if (layerNotRendered) {
-      handleGlobeUpdating(true);
-    }
-
-    handleLayerToggle(option.id);
-    handleLayerRendered(view, layerToggled, handleGlobeUpdating);
-
-    const isLayerActive = alreadyChecked[option.value];
-    if (isLayerActive) addLayerAnalyticsEvent({ slug: option.slug })
-    else removeLayerAnalyticsEvent({ slug: option.slug });
+    toggleWDPALayer(activeLayers, option, handleGlobeUpdating, view, map, handleLayerToggle);
   }
 
   const alreadyChecked = WDPALayers.reduce((acc, option) => ({
